[Storage] Move storage code to crate
