+++ crates/storage/Cargo.toml
@@ -2,7 +2,26 @@
 name = "storage"
 version = "0.1.0"
 edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+authors = [
+    "Furisto",
+    "Mrunal Patel <mrunalp@gmail.com>",
+    "Sascha Grunert <mail@saschagrunert.de>",
+    "utam0k <k0ma@utam0k.jp>",
+]
+documentation = "https://docs.rs/containrs"
+homepage = "https://github.com/containers/containrs"
+repository = "https://github.com/containers/containrs"
+license = "Apache-2.0"
+keywords = ["runtime", "kubernetes", "cri", "container", "pod"]
+categories = ["network-programming", "api-bindings"]
 
 [dependencies]
+anyhow = "1.0.43"
+getset = "0.1.1"
+log = { version = "0.4.14", features = ["serde", "std"] }
+rmp-serde = "0.15.5"
+serde = { version = "1.0.130", features = ["derive"] }
+sled = "0.34.6"
+
+[dev-dependencies]
+tempfile = "3.2.0"
